separate code: move loading of access list members to AccessListMembers component
