+++ frontend/resourceadm/components/AccessListDetails/AccessListDetail.test.tsx
@@ -33,17 +33,6 @@ const defaultProps: AccessListDetailProps = {
   backUrl: '/listadmin',
 };
 
-const membersPage2OrgNr = '987654321';
-const membersResults = {
-  data: [{ orgNr: '123456789', orgName: 'Skatteetaten', isSubParty: false }],
-  nextPage: 'http://at22-next-page',
-};
-
-const membersResultsPage2 = {
-  data: [{ orgNr: membersPage2OrgNr, orgName: 'Digitaliseringsdirektoratet', isSubParty: false }],
-  nextPage: '',
-};
-
 const updateAccessListMock = jest.fn().mockImplementation(() => Promise.resolve({}));
 
 describe('AccessListDetail', () => {
@@ -139,32 +128,6 @@ describe('AccessListDetail', () => {
     expect(
       screen.queryByText(textMock('resourceadm.listadmin_delete_list_header')),
     ).not.toBeInTheDocument();
-  });
-
-  it('should show more members when load more button is clicked', async () => {
-    const user = userEvent.setup();
-    const getAccessListMembersMock = jest
-      .fn()
-      .mockImplementationOnce(() => Promise.resolve(membersResults))
-      .mockImplementationOnce(() => Promise.resolve(membersResultsPage2));
-    renderAccessListDetail({}, { getAccessListMembers: getAccessListMembersMock });
-
-    await waitFor(() =>
-      screen.findByText(
-        textMock('resourceadm.listadmin_load_more', {
-          unit: textMock('resourceadm.listadmin_member_unit'),
-        }),
-      ),
-    );
-
-    const loadMoreButton = screen.getByText(
-      textMock('resourceadm.listadmin_load_more', {
-        unit: textMock('resourceadm.listadmin_member_unit'),
-      }),
-    );
-    await user.click(loadMoreButton);
-
-    expect(await screen.findByText(membersPage2OrgNr)).toBeInTheDocument();
   });
 });
