+++ frontend/resourceadm/components/AccessListMembers/AccessListMembers.test.tsx
@@ -15,6 +15,20 @@ const testEnv = 'tt02';
 const testListIdentifier = 'listid';
 const testMemberPartyId = '123456789';
 
+const membersPage2OrgNr = '987654321';
+const membersResults = {
+  data: [
+    { orgNr: testMemberPartyId, orgName: '', isSubParty: false },
+    { orgNr: '112233445', orgName: 'test', isSubParty: true },
+  ],
+  nextPage: 'http://at22-next-page',
+};
+
+const membersResultsPage2 = {
+  data: [{ orgNr: membersPage2OrgNr, orgName: 'Digitaliseringsdirektoratet', isSubParty: false }],
+  nextPage: '',
+};
+
 const defaultProps: AccessListMembersProps = {
   org: testOrg,
   env: testEnv,
@@ -24,19 +38,6 @@ const defaultProps: AccessListMembersProps = {
     name: 'Test-list',
     description: 'This is a description',
   },
-  members: [
-    {
-      orgNr: testMemberPartyId,
-      orgName: '',
-      isSubParty: false,
-    },
-    {
-      orgNr: '112233445',
-      orgName: 'test',
-      isSubParty: true,
-    },
-  ],
-  loadMoreButton: null,
   latestEtag: '',
   setLatestEtag: jest.fn(),
 };
@@ -44,20 +45,26 @@ const defaultProps: AccessListMembersProps = {
 describe('AccessListMembers', () => {
   afterEach(jest.clearAllMocks);
 
-  it('should show special party name if name is not found', () => {
-    renderAccessListMembers();
+  it('should show special party name if name is not found', async () => {
+    await renderAndWaitForData();
     expect(screen.getByText(textMock('resourceadm.listadmin_empty_name'))).toBeInTheDocument();
   });
 
-  it('should show message when list is empty', () => {
-    renderAccessListMembers({ members: [] });
-    expect(screen.getByText(textMock('resourceadm.listadmin_empty_list'))).toBeInTheDocument();
+  it('should show message when list is empty', async () => {
+    const getAccessListMembersMock = jest
+      .fn()
+      .mockImplementation(() => Promise.resolve({ data: [] }));
+    renderAccessListMembers({ getAccessListMembers: getAccessListMembersMock });
+
+    await waitFor(() => screen.findByText(textMock('resourceadm.listadmin_empty_list')));
   });
 
   it('should remove member from table when remove member button is clicked', async () => {
     const user = userEvent.setup();
     const removeAccessListMemberMock = jest.fn().mockImplementation(() => Promise.resolve({}));
-    renderAccessListMembers({}, { removeAccessListMember: removeAccessListMemberMock });
+    await renderAndWaitForData({
+      removeAccessListMember: removeAccessListMemberMock,
+    });
 
     const removeButtons = screen.getAllByText(textMock('resourceadm.listadmin_remove_from_list'));
     await user.click(removeButtons[0]);
@@ -71,21 +78,19 @@ describe('AccessListMembers', () => {
   it('should show new member in list after member is added', async () => {
     const user = userEvent.setup();
     const addAccessListMemberMock = jest.fn().mockImplementation(() => Promise.resolve({}));
+
     const searchResultText = 'Digdir';
     const searchResultOrgNr = '987654321';
-    renderAccessListMembers(
-      { loadMoreButton: <button></button> },
-      {
-        addAccessListMember: addAccessListMemberMock,
-        getParties: jest.fn().mockImplementation(() =>
-          Promise.resolve({
-            _embedded: {
-              enheter: [{ organisasjonsnummer: searchResultOrgNr, navn: searchResultText }],
-            },
-          }),
-        ),
-      },
-    );
+    await renderAndWaitForData({
+      addAccessListMember: addAccessListMemberMock,
+      getParties: jest.fn().mockImplementation(() =>
+        Promise.resolve({
+          _embedded: {
+            enheter: [{ organisasjonsnummer: searchResultOrgNr, navn: searchResultText }],
+          },
+        }),
+      ),
+    });
     const addMoreButton = screen.getByRole('button', {
       name: textMock('resourceadm.listadmin_search_add_more'),
     });
@@ -109,12 +114,9 @@ describe('AccessListMembers', () => {
   it('should show message when no parties are found', async () => {
     const user = userEvent.setup();
 
-    renderAccessListMembers(
-      {},
-      {
-        getParties: jest.fn().mockImplementation(() => Promise.resolve({})),
-      },
-    );
+    await renderAndWaitForData({
+      getParties: jest.fn().mockImplementation(() => Promise.resolve({})),
+    });
 
     const addMoreButton = screen.getByRole('button', {
       name: textMock('resourceadm.listadmin_search_add_more'),
@@ -130,12 +132,9 @@ describe('AccessListMembers', () => {
   it('should show message when no sub parties are found', async () => {
     const user = userEvent.setup();
 
-    renderAccessListMembers(
-      {},
-      {
-        getSubParties: jest.fn().mockImplementation(() => Promise.resolve({})),
-      },
-    );
+    await renderAndWaitForData({
+      getSubParties: jest.fn().mockImplementation(() => Promise.resolve({})),
+    });
 
     const addMoreButton = screen.getByRole('button', {
       name: textMock('resourceadm.listadmin_search_add_more'),
@@ -156,12 +155,9 @@ describe('AccessListMembers', () => {
   it('should show special organization from tenor when search for orgnr is not found', async () => {
     const user = userEvent.setup();
 
-    renderAccessListMembers(
-      {},
-      {
-        getParties: jest.fn().mockImplementation(() => Promise.resolve({})),
-      },
-    );
+    await renderAndWaitForData({
+      getParties: jest.fn().mockImplementation(() => Promise.resolve({})),
+    });
 
     const addMoreButton = screen.getByRole('button', {
       name: textMock('resourceadm.listadmin_search_add_more'),
@@ -179,21 +175,18 @@ describe('AccessListMembers', () => {
     const searchResultText = 'Digdir';
     const searchResultOrgNr = '987654321';
 
-    renderAccessListMembers(
-      {},
-      {
-        addAccessListMember: jest
-          .fn()
-          .mockImplementation(() => Promise.reject({ response: { data: { code: 'RR-00001' } } })),
-        getParties: jest.fn().mockImplementation(() =>
-          Promise.resolve({
-            _embedded: {
-              enheter: [{ organisasjonsnummer: searchResultOrgNr, navn: searchResultText }],
-            },
-          }),
-        ),
-      },
-    );
+    await renderAndWaitForData({
+      addAccessListMember: jest
+        .fn()
+        .mockImplementation(() => Promise.reject({ response: { data: { code: 'RR-00001' } } })),
+      getParties: jest.fn().mockImplementation(() =>
+        Promise.resolve({
+          _embedded: {
+            enheter: [{ organisasjonsnummer: searchResultOrgNr, navn: searchResultText }],
+          },
+        }),
+      ),
+    });
 
     const addMoreButton = screen.getByRole('button', {
       name: textMock('resourceadm.listadmin_search_add_more'),
@@ -216,21 +209,18 @@ describe('AccessListMembers', () => {
     const searchResultText = 'Digdir';
     const searchResultOrgNr = '987654321';
 
-    renderAccessListMembers(
-      {},
-      {
-        addAccessListMember: jest
-          .fn()
-          .mockImplementation(() => Promise.reject({ response: { status: 412, data: {} } })),
-        getParties: jest.fn().mockImplementation(() =>
-          Promise.resolve({
-            _embedded: {
-              enheter: [{ organisasjonsnummer: searchResultOrgNr, navn: searchResultText }],
-            },
-          }),
-        ),
-      },
-    );
+    await renderAndWaitForData({
+      addAccessListMember: jest
+        .fn()
+        .mockImplementation(() => Promise.reject({ response: { status: 412, data: {} } })),
+      getParties: jest.fn().mockImplementation(() =>
+        Promise.resolve({
+          _embedded: {
+            enheter: [{ organisasjonsnummer: searchResultOrgNr, navn: searchResultText }],
+          },
+        }),
+      ),
+    });
 
     const addMoreButton = screen.getByRole('button', {
       name: textMock('resourceadm.listadmin_search_add_more'),
@@ -253,14 +243,11 @@ describe('AccessListMembers', () => {
   it('should show error message if remove member request returns http status code 412', async () => {
     const user = userEvent.setup();
 
-    renderAccessListMembers(
-      {},
-      {
-        removeAccessListMember: jest
-          .fn()
-          .mockImplementation(() => Promise.reject({ response: { status: 412, data: {} } })),
-      },
-    );
+    await renderAndWaitForData({
+      removeAccessListMember: jest
+        .fn()
+        .mockImplementation(() => Promise.reject({ response: { status: 412, data: {} } })),
+    });
 
     const removeButtons = screen.getAllByText(textMock('resourceadm.listadmin_remove_from_list'));
     await user.click(removeButtons[0]);
@@ -274,6 +261,7 @@ describe('AccessListMembers', () => {
     const user = userEvent.setup();
     const nextPageUrl = 'brreg/next';
     const searchResultText = 'Digdir';
+
     const getSubPartiesMock = jest.fn().mockImplementation(() =>
       Promise.resolve({
         _embedded: {
@@ -287,12 +275,9 @@ describe('AccessListMembers', () => {
         },
       }),
     );
-    renderAccessListMembers(
-      {},
-      {
-        getSubParties: getSubPartiesMock,
-      },
-    );
+    await renderAndWaitForData({
+      getSubParties: getSubPartiesMock,
+    });
     const addMoreButton = screen.getByRole('button', {
       name: textMock('resourceadm.listadmin_search_add_more'),
     });
@@ -320,6 +305,7 @@ describe('AccessListMembers', () => {
     const user = userEvent.setup();
 
     const searchResultText = 'Digdir';
+
     const getPartiesMock = jest.fn().mockImplementation(() =>
       Promise.resolve({
         _embedded: {
@@ -332,12 +318,9 @@ describe('AccessListMembers', () => {
         },
       }),
     );
-    renderAccessListMembers(
-      {},
-      {
-        getParties: getPartiesMock,
-      },
-    );
+    await renderAndWaitForData({
+      getParties: getPartiesMock,
+    });
 
     const addMoreButton = screen.getByRole('button', {
       name: textMock('resourceadm.listadmin_search_add_more'),
@@ -359,22 +342,55 @@ describe('AccessListMembers', () => {
       ),
     ).toBeInTheDocument();
   });
+
+  it('should show more members when load more button is clicked', async () => {
+    const user = userEvent.setup();
+    const getAccessListMembersMock = jest
+      .fn()
+      .mockImplementationOnce(() => Promise.resolve(membersResults))
+      .mockImplementationOnce(() => Promise.resolve(membersResultsPage2));
+    await renderAndWaitForData({ getAccessListMembers: getAccessListMembersMock });
+
+    await waitFor(() =>
+      screen.findByText(
+        textMock('resourceadm.listadmin_load_more', {
+          unit: textMock('resourceadm.listadmin_member_unit'),
+        }),
+      ),
+    );
+
+    const loadMoreButton = screen.getByText(
+      textMock('resourceadm.listadmin_load_more', {
+        unit: textMock('resourceadm.listadmin_member_unit'),
+      }),
+    );
+    await user.click(loadMoreButton);
+
+    expect(await screen.findByText(membersPage2OrgNr)).toBeInTheDocument();
+  });
 });
 
-const renderAccessListMembers = (
-  props: Partial<AccessListMembersProps> = {},
-  queries: Partial<ServicesContextProps> = {},
-) => {
+const renderAccessListMembers = (queries: Partial<ServicesContextProps> = {}) => {
+  const defaultGetAccessListMembersMock = jest
+    .fn()
+    .mockImplementation(() => Promise.resolve(membersResults));
+
   const allQueries: ServicesContextProps = {
     ...queriesMock,
+    getAccessListMembers: defaultGetAccessListMembersMock,
     ...queries,
   };
 
-  return render(
+  render(
     <MemoryRouter>
       <ServicesContextProvider {...allQueries} client={createQueryClientMock()}>
-        <AccessListMembers {...defaultProps} {...props} />
+        <AccessListMembers {...defaultProps} />
       </ServicesContextProvider>
     </MemoryRouter>,
   );
+};
+
+const renderAndWaitForData = async (queries: Partial<ServicesContextProps> = {}) => {
+  renderAccessListMembers(queries);
+  await waitFor(() => screen.findByText(testMemberPartyId));
 };
